backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
pydantic==2.5.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0
python-jose[cryptography]==3.3.0
passlib==1.7.4
bcrypt==4.1.0
python-multipart==0.0.6
pytest==7.4.3
httpx==0.25.2
pytest-asyncio==0.21.1
alembic==1.13.0
requests==2.31.0